Fix imports after rename
