+++ src/features/schema-browser/utils/constraints.ts
@@ -1,5 +1,5 @@
 import { AsyncDuckDBConnectionPool } from '@features/duckdb-context/duckdb-connection-pool';
-import { ConstraintResult } from '@utils/duckdb/types';
+import { ConstraintResult } from '@utils/duckdb/models';
 
 import { escapeStringLiteral } from './sql-escape';
 import { vectorToArray } from './vector-extraction';
